requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
PyPDF2==3.0.1
google-genai==0.6.0
beautifulsoup4
requests
pytest>=7.4.0
pytest-mock>=3.12.0
